using useDelayToggleButtonState hook instead of a HOC
